separate preview with values component
